# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TransitOps API",
    "description": "Fleet operations platform \u2014 manage vehicles, drivers, trips, maintenance, fuel, expenses, and reports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/maintenance": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "summary": "List maintenance logs",
        "description": "Returns all maintenance logs.",
        "operationId": "list_maintenance_maintenance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceResponse"
                  },
                  "type": "array",
                  "title": "Response List Maintenance Maintenance Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Create maintenance record",
        "description": "Create a new maintenance record for a vehicle. Rule 9: automatically sets the vehicle status to 'In Shop'.",
        "operationId": "create_maintenance_maintenance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/maintenance/{log_id}": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Get maintenance log by ID",
        "operationId": "get_maintenance_maintenance__log_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Maintenance log not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maintenance/{log_id}/close": {
      "post": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Close a maintenance record",
        "description": "Close an open maintenance record. Rule 10: restores vehicle to 'Available' unless 'Retired'.",
        "operationId": "close_maintenance_maintenance__log_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MaintenanceClose"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Maintenance log not found"
          },
          "409": {
            "description": "Maintenance log already closed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fuel-expense/fuel": {
      "get": {
        "tags": [
          "Fuel & Expense"
        ],
        "summary": "List fuel logs",
        "description": "Returns all fuel logs. Optional filter by vehicle_id.",
        "operationId": "list_fuel_logs_fuel_expense_fuel_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuelLogResponse"
                  },
                  "title": "Response List Fuel Logs Fuel Expense Fuel Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Fuel & Expense"
        ],
        "summary": "Create a fuel log",
        "description": "Record a fuel log entry for a vehicle.",
        "operationId": "create_fuel_log_fuel_expense_fuel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuelLogCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelLogResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fuel-expense/expenses": {
      "get": {
        "tags": [
          "Fuel & Expense"
        ],
        "summary": "List expenses",
        "description": "Returns all expenses. Optional filter by vehicle_id.",
        "operationId": "list_expenses_fuel_expense_expenses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseResponse"
                  },
                  "title": "Response List Expenses Fuel Expense Expenses Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Fuel & Expense"
        ],
        "summary": "Create an expense",
        "description": "Record an expense entry for a vehicle.",
        "operationId": "create_expense_fuel_expense_expenses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fuel-expense/operational-cost/{vehicle_id}": {
      "get": {
        "tags": [
          "Fuel & Expense"
        ],
        "summary": "Get total operational cost for a vehicle",
        "description": "Computes total operational cost = Fuel + Maintenance + Expenses for a vehicle.",
        "operationId": "get_operational_cost_fuel_expense_operational_cost__vehicle_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExpenseCategory": {
        "type": "string",
        "enum": [
          "Toll",
          "Fine",
          "Parking",
          "Other"
        ],
        "title": "ExpenseCategory"
      },
      "ExpenseCreate": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "examples": [
              "uuid-of-vehicle"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/ExpenseCategory",
            "examples": [
              "Toll"
            ]
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "examples": [
              25.0
            ]
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "examples": [
              "2026-01-15"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "examples": [
              "Highway toll \u2014 Route 66"
            ]
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "category",
          "amount",
          "date"
        ],
        "title": "ExpenseCreate"
      },
      "ExpenseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "category": {
            "$ref": "#/components/schemas/ExpenseCategory"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "category",
          "amount",
          "date",
          "notes"
        ],
        "title": "ExpenseResponse"
      },
...
      "FuelLogCreate": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "examples": [
              "uuid-of-vehicle"
            ]
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id",
            "examples": [
              "uuid-of-trip"
            ]
          },
          "liters": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Liters",
            "examples": [
              45.0
            ]
          },
          "cost": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost",
            "examples": [
              67.5
            ]
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "examples": [
              "2026-01-15"
            ]
          },
          "odometer_km": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer Km",
            "examples": [
              15350.0
            ]
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "liters",
          "cost",
          "date"
        ],
        "title": "FuelLogCreate"
      },
      "FuelLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "liters": {
            "type": "number",
            "title": "Liters"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "odometer_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer Km"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "trip_id",
          "liters",
          "cost",
          "date",
          "odometer_km"
        ],
        "title": "FuelLogResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MaintenanceClose": {
        "properties": {
          "cost": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost",
            "examples": [
              275.0
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Completed oil change and filter replacement"
            ]
          }
        },
        "type": "object",
        "title": "MaintenanceClose"
      },
      "MaintenanceCreate": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "examples": [
              "uuid-of-vehicle"
            ]
          },
          "service_type": {
            "type": "string",
            "title": "Service Type",
            "examples": [
              "Oil Change"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Routine 10k km oil change"
            ]
          },
          "cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cost",
            "default": 0.0,
            "examples": [
              250.0
            ]
          },
          "odometer_at_service_km": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer At Service Km",
            "examples": [
              15000.0
            ]
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "service_type"
        ],
        "title": "MaintenanceCreate"
      },
      "MaintenanceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "odometer_at_service_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer At Service Km"
          },
          "status": {
            "$ref": "#/components/schemas/MaintenanceStatus"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "service_type",
          "description",
          "cost",
          "odometer_at_service_km",
          "status",
          "opened_at",
          "closed_at"
        ],
        "title": "MaintenanceResponse"
      },
      "MaintenanceStatus": {
        "type": "string",
        "enum": [
          "Open",
          "Closed"
        ],
        "title": "MaintenanceStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}